# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FAM Backend API",
    "version": "1.0.0",
    "description": "Family event management platform — three-role auth (user/admin/organizer), family tree, QR join, notifications."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RelationshipObject": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "eventId": { "type": "string" },
          "person1": { "type": "string" },
          "person2": { "type": "string" },
          "relationType": { "type": "string", "example": "parent" },
          "familySide": { "type": "string", "enum": ["groom", "bride", "common"] },
          "isValidated": { "type": "boolean" },
          "marriageDate": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/tree/generate/{eventId}": {
      "get": {
        "summary": "Get the current family tree. treeVisibility on the event controls access.",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "familySide",
            "schema": { "type": "string", "enum": ["groom", "bride", "common"] },
            "description": "For wedding/anniversary events — returns only that sub-tree"
          }
        ],
        "responses": {
          "200": {
            "description": "Tree with nodes[], edges[], and mermaidCode",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "tree": {
                      "type": "object",
                      "properties": {
                        "nodes": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "label": { "type": "string" },
                              "data": {
                                "type": "object",
                                "properties": {
                                  "username": { "type": "string" },
                                  "profileImage": { "type": "string" },
                                  "isDeceased": { "type": "boolean" },
                                  "deathYear": { "type": "integer", "nullable": true }
                                }
                              }
                            }
                          }
                        },
                        "edges": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "source": { "type": "string" },
                              "target": { "type": "string" },
                              "label": { "type": "string", "example": "parent" }
                            }
                          }
                        },
                        "mermaidCode": { "type": "string", "example": "flowchart TD\n  A[Alice] --> B[Bob]" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tree/history/{eventId}": {
      "get": {
        "summary": "Tree snapshot history. Users must be participants; organizers must be assigned to the event.",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "version", "schema": { "type": "integer" }, "description": "Specific snapshot version number" },
          { "in": "query", "name": "familySide", "schema": { "type": "string", "enum": ["groom", "bride", "common"] } }
        ],
        "responses": {
          "200": { "description": "Array of tree snapshots ordered by date" }
        }
      }
    },
    "/api/tree/path/{eventId}": {
      "get": {
        "summary": "Discover how two people are related",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "fromUserId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "toUserId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Relationship path between two users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "found": { "type": "boolean" },
                    "path": { "type": "array", "items": { "type": "string" }, "example": ["userId1", "userId2", "userId3"] },
                    "length": { "type": "integer", "example": 2 }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tree/add-relation": {
      "post": {
        "summary": "Add a family relationship. Users added by regular users start with isValidated=false; organizer/admin skip validation.",
        "description": "Validation rules: no duplicate edges, no circular ancestry, parent must be ≥12 years older than child, non-main users can add at most 4 relationships. For wedding/anniversary events familySide is required.",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId", "person1Id", "person2Id", "relationType"],
                "properties": {
                  "eventId": { "type": "string" },
                  "person1Id": { "type": "string", "description": "MongoDB ObjectId of first person" },
                  "person2Id": { "type": "string", "description": "MongoDB ObjectId of second person" },
                  "relationType": {
                    "type": "string",
                    "enum": ["parent", "child", "spouse", "sibling", "friend", "step-parent", "step-child", "step-sibling", "grandparent", "grandchild", "cousin", "uncle", "aunt", "nephew", "niece", "father-in-law", "mother-in-law", "brother-in-law", "sister-in-law"],
                    "example": "parent"
                  },
                  "familySide": {
                    "type": "string",
                    "enum": ["groom", "bride", "common"],
                    "description": "Required for wedding/anniversary events"
                  },
                  "marriageDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Optional, only meaningful for relationType=spouse",
                    "example": "2010-05-20"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Relationship added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Relationship added successfully" },
                    "relationship": { "$ref": "#/components/schemas/RelationshipObject" },
                    "tree": { "type": "object", "description": "Updated tree with nodes, edges, mermaidCode" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failure — includes errors array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "message": { "type": "string", "example": "Validation failed" },
                    "errors": { "type": "array", "items": { "type": "string" }, "example": ["This relationship already exists", "Parent must be older than child"] }
                  }
                }
              }
            }
          },
          "403": { "description": "Must be an approved participant to add relationships" }
        }
      }
    },
    "/api/tree/remove-relation/{id}": {
      "delete": {
        "summary": "Remove a relationship (admin/organizer only)",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId"],
                "properties": {
                  "eventId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Relationship removed, returns updated tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Relationship removed successfully" },
                    "tree": { "type": "object" }
                  }
                }
              }
            }
          },
          "403": { "description": "Not authorized or relationship not found" }
        }
      }
    },
    "/api/tree/update-relation/{id}": {
      "patch": {
        "summary": "Update a relationship (admin/organizer only)",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "relationType": { "type": "string", "example": "sibling" },
                  "familySide": { "type": "string", "enum": ["groom", "bride", "common"] },
                  "marriageDate": { "type": "string", "format": "date", "example": "2010-05-20" },
                  "isValidated": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Relationship updated" }
        }
      }
    },
    "/api/tree/set-main-person": {
      "post": {
        "summary": "Set the central person for a common event. This person is exempt from the 4-person add limit. (admin/organizer only)",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId", "userId"],
                "properties": {
                  "eventId": { "type": "string" },
                  "userId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Main person set" }
        }
      }
    },
    "/api/tree/set-wedding-couple": {
      "post": {
        "summary": "Set groom and bride for a wedding event. Creates a spouse relationship and sets event.treeType=wedding. (admin/organizer only)",
        "tags": ["Tree"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId", "groomId", "brideId"],
                "properties": {
                  "eventId": { "type": "string" },
                  "groomId": { "type": "string" },
                  "brideId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Wedding couple configured" }
        }
      }
    },
...
  },
...
}